[refactor] service 관련 구조 리팩터
